test(collection): hoist NC_FIXTURE + the duplicated three_files fixtures (Theme D)
Add collection/conftest.py with the shared NC_FIXTURE path constant and the two
identical three_files fixtures (three_files for the 4x5 fill-i pair, three_files_ramp
for the 3x4 fill-(i+1) pair); the test_reductions.py solo variant stays. No test
logic changes.
