+++ tests/dataset/collection/test_meta.py
@@ -37,23 +37,6 @@ def template_file(tmp_path):
     return path
 
 
-@pytest.fixture
-def three_files(tmp_path):
-    paths = []
-    for i in range(3):
-        arr = np.full((4, 5), i, dtype=np.float32)
-        ds = Dataset.create_from_array(
-            arr,
-            top_left_corner=(0.0, 4.0),
-            cell_size=1.0,
-            epsg=4326,
-        )
-        p = str(tmp_path / f"f{i}.tif")
-        ds.to_file(p)
-        paths.append(p)
-    return paths
-
-
 class TestMetaProperty:
     def test_meta_exposed_on_existing_constructor(self, template_file):
         src = Dataset.read_file(template_file)
